requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
